+++ src/Redis/Command/Profile.php
@@ -21,46 +21,72 @@ declare(strict_types=1);
 
 namespace MacFJA\RediSearch\Redis\Command;
 
-use MacFJA\RediSearch\Redis\Command\Option\CustomValidatorOption;
+use MacFJA\RediSearch\Redis\Command\Option\CustomValidatorOption as CV;
 use MacFJA\RediSearch\Redis\Command\Option\FlagOption;
 use MacFJA\RediSearch\Redis\Command\Option\NamelessOption;
+use MacFJA\RediSearch\Redis\Command\ProfileCommand\QueryOption;
 
-class ConfigSet extends AbstractCommand
+class Profile extends AbstractCommand
 {
     public function __construct(string $rediSearchVersion = self::MIN_IMPLEMENTED_VERSION)
     {
         parent::__construct([
-            'action' => new FlagOption('SET', true, '>=2.0.0'),
-            'options' => [
-                CustomValidatorOption::allowedValues(new NamelessOption(null, '>=2.0.0 <2.0.6'), ['NOGC', 'MAXEXPANSIONS', 'TIMEOUT', 'ON_TIMEOUT', 'MIN_PHONETIC_TERM_LEN']),
-                CustomValidatorOption::allowedValues(new NamelessOption(null, '>=2.0.6 <2.4.3'), ['NOGC', 'MINPREFIX', 'MAXEXPANSIONS', 'MAXFILTEREXPANSION', 'TIMEOUT', 'ON_TIMEOUT', 'MIN_PHONETIC_TERM_LEN']),
-                CustomValidatorOption::allowedValues(new NamelessOption(null, '>=2.4.3'), ['NOGC', 'MINPREFIX', 'MAXEXPANSIONS', 'TIMEOUT', 'ON_TIMEOUT', 'MIN_PHONETIC_TERM_LEN', 'DEFAULT_DIALECT']),
-            ],
-            'value' => new NamelessOption(null, '>=2.0.0'),
+            'index' => new NamelessOption(null, '>=2.2.0'),
+            'type' => CV::allowedValues(new NamelessOption(null, '>=2.2.0'), ['SEARCH', 'AGGREGATE']),
+            'limited' => new FlagOption('LIMITED', false, '>=2.2.0'),
+            'query' => new QueryOption('>=2.2.0'),
         ], $rediSearchVersion);
     }
 
-    /**
-     * @param mixed $value
-     *
-     * @return $this
-     */
-    public function setConfig(string $name, $value): self
+    public function setIndex(string $index): self
     {
-        $this->options['options'][0]->setValue($name);
-        $this->options['options'][1]->setValue($name);
-        $this->options['value']->setValue($value);
+        $this->options['index']->setValue($index);
+
+        return $this;
+    }
+
+    /**
+     * @param Aggregate|Search $query
+     */
+    public function setQuery($query): self
+    {
+        /** @var QueryOption $queryOption */
+        $queryOption = $this->options['query'];
+        $queryOption->setCommand($query);
+        $this->options['type']->setValue($queryOption->isSearch() ? 'SEARCH' : 'AGGREGATE');
+        $this->options['index']->setValue($queryOption->getIndex());
+
+        return $this;
+    }
+
+    public function setTypeAggregate(): self
+    {
+        $this->options['type']->setValue('AGGREGATE');
+
+        return $this;
+    }
+
+    public function setTypeSearch(): self
+    {
+        $this->options['type']->setValue('SEARCH');
+
+        return $this;
+    }
+
+    public function setLimited(bool $active = true): self
+    {
+        $this->options['limited']->setActive($active);
 
         return $this;
     }
 
     public function getId(): string
     {
-        return 'FT.CONFIG';
+        return 'FT.PROFILE';
     }
 
     protected function getRequiredOptions(): array
     {
-        return ['action', 'options', 'value'];
+        return ['index', 'type', 'query'];
     }
 }