+++ src/Redis/Command/Search.php
@@ -30,6 +30,7 @@ use MacFJA\RediSearch\Redis\Command\Option\CustomValidatorOption as CV;
 use MacFJA\RediSearch\Redis\Command\Option\FlagOption;
 use MacFJA\RediSearch\Redis\Command\Option\NamedOption;
 use MacFJA\RediSearch\Redis\Command\Option\NamelessOption;
+use MacFJA\RediSearch\Redis\Command\Option\NotEmptyOption;
 use MacFJA\RediSearch\Redis\Command\Option\NumberedOption;
 use MacFJA\RediSearch\Redis\Command\SearchCommand\FilterOption;
 use MacFJA\RediSearch\Redis\Command\SearchCommand\GeoFilterOption;
@@ -42,7 +43,6 @@ use MacFJA\RediSearch\Redis\Response\PaginatedResponse;
 use MacFJA\RediSearch\Redis\Response\SearchResponseItem;
 
 /**
- * @method PaginatedResponse<SearchResponseItem> parseResponse(mixed $data)
  * @SuppressWarnings(PHPMD.CouplingBetweenObjects)
  */
 class Search extends AbstractCommand implements PaginatedCommand
@@ -76,8 +76,10 @@ class Search extends AbstractCommand implements PaginatedCommand
                 'scorer' => new NamedOption('SCORER', null, '>=2.0.0'),
                 'explainscore' => new FlagOption('EXPLAINSCORE', false, '>=2.0.0'),
                 'payload' => new NamedOption('PAYLOAD', null, '>=2.0.0'),
+                'params' => new NotEmptyOption(new NumberedOption('PARAMS', null, '>=2.4.0')),
                 'sortby' => new SortByOption(),
                 'limit' => new LimitOption(),
+                'dialect' => CV::isNumeric(new NamedOption('DIALECT', null, '>=2.4.3')),
             ],
             $rediSearchVersion
         );
@@ -88,6 +90,11 @@ class Search extends AbstractCommand implements PaginatedCommand
         $this->options['index']->setValue($index);
 
         return $this;
+    }
+
+    public function getIndex(): string
+    {
+        return $this->options['index']->getValue();
     }
 
     public function setQuery(string $query): self
@@ -281,6 +288,21 @@ class Search extends AbstractCommand implements PaginatedCommand
         return $this;
     }
 
+    /**
+     * @param int|scalar|string $value
+     *
+     * @return $this
+     */
+    public function addParam(string $name, $value): self
+    {
+        $params = $this->options['params']->getArguments() ?? [];
+        $params[] = $name;
+        $params[] = (string) $value;
+        $this->options['params']->setArguments($params);
+
+        return $this;
+    }
+
     public function getId(): string
     {
         return 'FT.SEARCH';
@@ -302,10 +324,18 @@ class Search extends AbstractCommand implements PaginatedCommand
         return $limit->getSize();
     }
 
+    public function setDialect(int $version): self
+    {
+        $this->options['dialect']->setValue($version);
+
+        return $this;
+    }
+
     /**
      * @param mixed $data
      *
      * @return PaginatedResponse
+     * @psalm-return PaginatedResponse<SearchResponseItem>
      *
      * @SuppressWarnings(PHPMD.CyclomaticComplexity)
      * @SuppressWarnings(PHPMD.NPathComplexity)