+++ tests/Redis/Command/AggregateTest.php
@@ -121,9 +121,51 @@ class AggregateTest extends TestCase
         static::assertSame(10, $command->getSize());
     }
 
+    public function testLoadAll(): void
+    {
+        $command = new Aggregate('2.0.12');
+        $command
+            ->setIndex('idx')
+            ->setQuery('@text1:"hello world"')
+            ->setLoadAll()
+        ;
+
+        static::assertSame([
+            'idx',
+            '@text1:"hello world"',
+        ], $command->getArguments());
+
+        $command = new Aggregate('2.0.13');
+        $command
+            ->setIndex('idx')
+            ->setQuery('@text1:"hello world"')
+            ->setLoadAll()
+        ;
+
+        static::assertSame([
+            'idx',
+            '@text1:"hello world"',
+            'LOAD', 'ALL',
+        ], $command->getArguments());
+
+        $command = new Aggregate('2.0.13');
+        $command
+            ->setIndex('idx')
+            ->setQuery('@text1:"hello world"')
+            ->setLoad('foo', 'bar')
+            ->setLoadAll()
+        ;
+
+        static::assertSame([
+            'idx',
+            '@text1:"hello world"',
+            'LOAD', 'ALL',
+        ], $command->getArguments());
+    }
+
     public function testFullOption(): void
     {
-        $command = new Aggregate();
+        $command = new Aggregate(Aggregate::MAX_IMPLEMENTED_VERSION);
         $command
             ->setIndex('idx')
             ->setQuery('@text1:"hello world"')
@@ -137,6 +179,7 @@ class AggregateTest extends TestCase
             ->setLimit(12, 35)
             ->addApply('@timestamp - (@timestamp % 86400)', 'day')
             ->setWithCursor(20, 30)
+            ->setDialect(2)
         ;
 
         static::assertSame([
@@ -150,6 +193,7 @@ class AggregateTest extends TestCase
             'LIMIT', 12, 35,
             'FILTER', "@name=='foo' && @age < 20",
             'WITHCURSOR', 'COUNT', 20, 'MAXIDLE', 30,
+            'DIALECT', 2,
         ], $command->getArguments());
     }