+++ tests/Redis/Command/ExplainTest.php
@@ -30,6 +30,9 @@ use PHPUnit\Framework\TestCase;
  * @covers \MacFJA\RediSearch\Redis\Command\Explain
  *
  * @uses \MacFJA\RediSearch\Redis\Command\Option\AbstractCommandOption
+ * @uses \MacFJA\RediSearch\Redis\Command\Option\CustomValidatorOption
+ * @uses \MacFJA\RediSearch\Redis\Command\Option\DecoratedOptionTrait
+ * @uses \MacFJA\RediSearch\Redis\Command\Option\NamedOption
  * @uses \MacFJA\RediSearch\Redis\Command\Option\NamelessOption
  *
  * @internal
@@ -44,15 +47,17 @@ class ExplainTest extends TestCase
 
     public function testFullOption(): void
     {
-        $command = new Explain();
+        $command = new Explain(Explain::MAX_IMPLEMENTED_VERSION);
         $command
             ->setIndex('idx')
             ->setQuery('@text1:"hello world"')
+            ->setDialect(2)
         ;
 
         static::assertSame([
             'idx',
             '@text1:"hello world"',
+            'DIALECT', 2,
         ], $command->getArguments());
     }
 }