+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Documentation on paginated responses
 - Throw custom exception is client is missing
 - Add compatibility for Predis version 2.x
+- `LOAD ALL` option on Aggregate command (RediSearch `2.0.13`)
+- Profile command (RediSearch `2.2.0`) ([Issue#4])
+- Vector feature (RediSearch `2.4.0` / `2.4.2`)
+- Add `DIALECT` option (RediSearch `2.4.3`) ([Issue#35])
 - (dev) Add `make clean` to remove all generated files
 - (dev) Add code coverage to `\MacFJA\RediSearch\Index`
 - (dev) Add integration test for document insertion
@@ -222,6 +226,7 @@ First version
 [1.0.0]: https://github.com/MacFJA/php-redisearch/releases/tag/1.0.0
 
 [Issue#2]: https://github.com/MacFJA/php-redisearch/issues/2
+[Issue#4]: https://github.com/MacFJA/php-redisearch/issues/4
 [Issue#5]: https://github.com/MacFJA/php-redisearch/issues/5
 [Issue#6]: https://github.com/MacFJA/php-redisearch/issues/6
 [Issue#9]: https://github.com/MacFJA/php-redisearch/issues/9
@@ -230,6 +235,7 @@ First version
 [Issue#12]: https://github.com/MacFJA/php-redisearch/issues/12
 [Issue#16]: https://github.com/MacFJA/php-redisearch/issues/16
 [Issue#26]: https://github.com/MacFJA/php-redisearch/issues/26
+[Issue#35]: https://github.com/MacFJA/php-redisearch/issues/35
 [Issue#38]: https://github.com/MacFJA/php-redisearch/issues/38
 [Issue#39]: https://github.com/MacFJA/php-redisearch/issues/39
 [Issue#46]: https://github.com/MacFJA/php-redisearch/issues/46